todo actions refactored and types extracted to centeral interfaces folder
