+++ frontend-service/src/interfaces/todo.d.ts
@@ -1,0 +1,6 @@
+import { Document } from 'mongoose';
+
+interface ITodo extends Document {
+	body: string;
+	completed: boolean;
+}
