+++ frontend-service/src/interfaces/user.d.ts
@@ -1,10 +1,5 @@
 import { Document } from 'mongoose';
 
-interface ITodo extends Document {
-	body: string;
-	completed: boolean;
-}
-
 interface IUser extends Document {
 	email: string;
 	password: string;
